Refactoring Allocators.
- Implementation of abstract allocators moved to an specific folder;
- Excluded the `DepSolverQAllocator` (`Dynprog` and `Simple` also
refactored);
- Legacy `BoundedSIDepSolver` removed;
- `QbitAllocator::executeAllocation` renamed to `QbitAllocator::allocate`;
- `QbitAllocator::allocate` returns the initial mapping only.
