+++ include/enfield/Transform/Allocators/IBMQAllocator.h
@@ -1,13 +1,13 @@
 #ifndef __EFD_IBM_QALLOCATOR_H__
 #define __EFD_IBM_QALLOCATOR_H__
 
-#include "enfield/Transform/Allocators/QbitAllocator.h"
+#include "enfield/Transform/Allocators/StdSolutionQAllocator.h"
 #include "enfield/Transform/LayersBuilderPass.h"
 #include "enfield/Support/Defs.h"
 
 namespace efd {
     /// \brief Port of IBM's allocator.
-    class IBMQAllocator : public QbitAllocator {
+    class IBMQAllocator : public StdSolutionQAllocator {
         public:
             typedef IBMQAllocator* Ref;
             typedef std::unique_ptr<IBMQAllocator> uRef;
@@ -31,7 +31,7 @@ namespace efd {
         public:
             IBMQAllocator(ArchGraph::sRef archGraph);
 
-            Solution executeAllocation(QModule::Ref qmod) override;
+            Solution buildStdSolution(QModule::Ref qmod) override;
             static uRef Create(ArchGraph::sRef archGraph);
     };
 }
